requirement.txt
Pillow
pyscreenshot
pytesseract
